Type Fulgora's f32 constants, taking 13 fields to bit-exact (#273)
Fulgora's chain held f64 literals where the game holds f32, plus a few
intermediates rounded once at the end rather than per operation. Eight fixes
are taken here and 13 named fields go from inexact to bit-exact against the
game: 31/67 -> 44/67 fully exact.

| fix | field | before | after |
| --- | --- | --- | --- |
| wobbleMask +0.6 typed | fulgora_wobble_mask | 96 | 101 |
| rock 0.33 typed | fulgora_rock | 84 | 101 |
| dunes 0.66 typed | fulgora_dunes | 26 | 101 |
| ruinsWalls 0.66 typed | fulgora_ruins_walls | 19 | 101 |
| natural 0.85 typed | fulgora_natural | 51 | 101 |
| sprawlPyramids per-op | fulgora_sprawl_pyramids | 99 | 101 |
| basisOil coords per-op | fulgora_basis_oil | 97 | 101 |
| tileRuinWalls per-op | fulgora_tile_ruin_walls | 75 | 101 |

basis, wx, wy, pyramids and pyramids_banding reach 101/101 with them.

Every candidate was applied to the real tree, scored, and reverted, in a loop
that re-baselined after each accept. One was taken only when its own field
reached a full 101/101 at a residual of exactly 0 - the standard f32.ts sets.
Twelve candidates that merely improved were rejected and tabled in #279.

Three findings the issue did not predict:

- fulgora_natural reaches 101/101, not the 99/101 predicted. That prediction
  was measured against a fixed baseline, and natural reads basis, which was
  itself only 98/101 until wobbleMask was typed. A one-at-a-time sweep
  under-reports a DAG.
- sprawlPyramids gets WORSE from typing its constants alone, 99 -> 97, and
  reaches 101/101 only with every intermediate narrowed. Which case applies
  depends on arity, not just on the literal.
- Tier 2 caught a port divergence tier 1 could not. The first Rust mirror of
  tileRuinWalls bound `let rw = ruins_walls as f32` and reused it, which is a
  different computation because these fields are held as f64 that have not been
  rounded to f32. Both ports scored 101/101 at every fixture position and
  disagreed off them; only the 676-point fold found it.

What did not move: the tile argmax is 4,915 of 5,057 before and after, with the
same 7 land/ocean and 11 shallow/deep misses, so those really are
boundary-exclusive. The terrain preview went 34,976 -> 34,977 differing pixels
of 1,048,576 - one pixel worse. This class of fix buys bit-exactness on named
fields, not a better picture; the image is dominated by the mix_* chain #273
could not reach.

Two frozen counts dropped by one and are recorded rather than smoothed:
fulgora_pre_elevation 44 -> 43 and fulgora_tile_ruin_machinery 95 -> 94. Both
are still-inexact deep composites, so inputs moving closer to the game
reshuffles which positions happen to round the same way.

The rest of the chain is blocked on #279: starting_spot_at_angle evaluates in
f64, and narrowing it takes both starting cones to 101/101 but regresses
vulcanus_starting_calcite, because it is shared with Vulcanus and its call
sites read the un-narrowed sliderRescale of #270.

pnpm run verify rc=0, 2057 passed, 4 skipped. engine.wasm rebuilt and committed.

+++ crates/fmw-noise/src/expressions/fulgora_shared.rs
@@ -236,7 +236,14 @@ impl FulgoraShared {
         // "We usually want a lot of wobble or none at all, so influence has a
         // high output scale and then we clamp it." The +0.6 biases most of the
         // map to fully on rather than centring the mask.
-        let wobble_mask = clamp(wobble_influence + 0.6, 0.0, 1.0);
+        //
+        // `0.6f32` is case 2 from the `eval` module docs - the engine holds the
+        // literal as 0.60000002384185791016 and the f64 one is
+        // 0.59999999999999997780. 96/101 exact before, **101/101 at a residual
+        // of exactly 0** after, and it is what takes `wx`, `wy`,
+        // `fulgora_basis`, `fulgora_pyramids` and `fulgora_pyramids_banding` to
+        // 101/101 as well. See #273.
+        let wobble_mask = clamp(wobble_influence + f64::from(0.6f32), 0.0, 1.0);
 
         // Offset the grid by half a cell so spawn sits in the MIDDLE of a cell
         // rather than on a corner where four islands meet.
